backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
passlib[bcrypt]
aiofiles
python-dotenv
Email-Validator
